requirements.txt
APScheduler==3.6.3
asgiref==3.2.3
bcrypt==3.1.7
boto3==1.12.3
botocore==1.15.3
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
Django==3.0.2
django-extensions==2.2.9
django-rest-framework==0.1.0
django-widget-tweaks==1.4.5
djangorestframework==3.11.0
docker==4.1.0
docutils==0.15.2
getmac==0.8.2
idna==2.8
ifaddr==0.1.6
jmespath==0.9.4
paramiko==2.7.1
pycparser==2.19
PyNaCl==1.3.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.0
tzlocal==2.0.0
urllib3==1.25.8
websocket-client==0.57.0
